requirements.txt
nextcord==2.6.0
python-dotenv==1.0.1
wakeonlan==3.1.0